src/requirement.txt
cycler
kiwisolver
matplotlib
numpy
opencv-python
pyparsing
python-dateutil
six
mecab-python3
